0.1.1 - temp sensors now work!

+++ CPP_flight_computer_program/include/sensor_ms5611.h
@@ -1,3 +1,3 @@
 bool init_MS5611();
-double get_ms5611_temp();
-double get_ms5611_press();
+float get_ms5611_temp();
+float get_ms5611_press();
